+++ tests/test_versions.py
@@ -3,9 +3,11 @@ import pytest
 from ccbr_actions.versions import (
     get_releases,
     get_latest_release_tag,
+    get_latest_release_hash,
     match_semver,
     check_version_increments_by_one,
     get_major_minor_version,
+    is_ancestor,
 )
 
 
@@ -15,6 +17,15 @@ def test_get_releases():
 
 def test_get_latest_release_tag():
     assert match_semver(get_latest_release_tag(), with_leading_v=True)
+
+
+def test_get_latest_release_hash():
+    assert all(
+        [
+            len(get_latest_release_hash()) > 7,
+            get_latest_release_hash(args="--repo CCBR/CCBR_NextflowTemplate") == "",
+        ]
+    )
 
 
 def test_version_increment():
@@ -51,6 +62,9 @@ def test_version_increment_error():
     messages.append(
         "Tag v10 does not match semantic versioning guidelines." in str(exc_info.value)
     )
+    with pytest.raises(ValueError) as exc_info:
+        check_version_increments_by_one("v1", "10", with_leading_v=True)
+    messages.append("The tag does not start with 'v'." in str(exc_info.value))
     assert all(messages)
 
 
@@ -61,5 +75,15 @@ def test_get_major_minor():
             get_major_minor_version("2.1.3-alpha") == "2.1",
             get_major_minor_version("v1.0.0", with_leading_v=True) == "v1.0",
             get_major_minor_version("invalid_version") == None,
+        ]
+    )
+
+
+def test_is_ancestor():
+    assert all(
+        [
+            is_ancestor("v0.1.0", "v0.1.1"),
+            is_ancestor("d620b61", "6cf677a"),
+            is_ancestor("d620b61\n", "\n6cf677a"),
         ]
     )